requirements.txt
appdirs==1.4.0
click==6.7
Flask==0.12
Flask-HTTPAuth==3.2.1
Flask-Redis==0.3.0
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
mockredispy==2.9.3
packaging==16.8
psycopg2==2.6.2
py==1.4.32
pyparsing==2.1.10
pytest==3.0.6
redis==2.10.5
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
Werkzeug==0.11.15
wheel==0.24.0
